Refactor: Move core files to src directory and update paths
- I moved App.tsx and index.tsx into the src/ directory.
- I renamed src/index.tsx to src/main.tsx.
- I created src/index.css for Tailwind directives and global styles.
- I updated index.html to point to src/main.tsx and src/index.css.
- I removed Tailwind CDN and inline config from index.html in favor of local CSS.
- I updated imports in src/main.tsx to reflect new file locations and include CSS.
