+++ index.html
@@ -3,68 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale-1.0" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Optiverse AI Agency</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-    <script>
-      tailwind.config = {
-        darkMode: 'class', // or 'media' if you prefer
-        theme: {
-          extend: {
-            colors: {
-              emerald: {
-                500: '#10B981',
-                600: '#059669',
-                900: '#064E3B', // Darker emerald for gradients if needed
-              },
-              purple: {
-                500: '#A855F7',
-                600: '#7C3AED',
-                900: '#3730A3',
-              },
-              pink: { // For gradients as used in hero title
-                500: '#EC4899',
-              },
-              gray: {
-                900: '#111827', 
-                800: '#1F2937',
-                700: '#374151',
-                600: '#4B5563',
-                500: '#6B7280', 
-                400: '#9CA3AF',
-                300: '#D1D5DB',
-                200: '#E5E7EB', // For light mode borders
-              },
-              yellow: { // For dark mode toggle icon and testimonial stars
-                400: '#FACC15',
-              },
-              green: { // For form success message
-                500: '#22C55E',
-              }
-            },
-            animation: {
-              blob: "blob 7s infinite",
-            },
-            keyframes: {
-              blob: {
-                "0%": {
-                  transform: "translate(0px, 0px) scale(1)",
-                },
-                "33%": {
-                  transform: "translate(30px, -50px) scale(1.1)",
-                },
-                "66%": {
-                  transform: "translate(-20px, 20px) scale(0.9)",
-                },
-                "100%": {
-                  transform: "translate(0px, 0px) scale(1)", // Corrected 'tranlate' to 'translate'
-                },
-              },
-            },
-          }
-        }
-      }
-    </script>
+    <link rel="stylesheet" href="/src/index.css">
   <style>
     html {
       scroll-behavior: smooth;
@@ -120,7 +61,6 @@
 </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/index.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
-</html><link rel="stylesheet" href="index.css">
-<script src="index.tsx" type="module"></script>
+</html>
