+++ src/main.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App'; // Path updated as App.tsx is now in the same src/ directory
+import App from './App.tsx'; // Path updated as App.tsx is now in the same src/ directory
+import './index.css';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {
